Restructuring component folders
